+++ pytgbot/bot/synchronous.py
@@ -6,6 +6,7 @@ from datetime import timedelta, datetime
 from DictObject import DictObject
 from luckydonaldUtils.logger import logging
 from luckydonaldUtils.encoding import unicode_type, to_unicode as u, to_native as n
+from luckydonaldUtils.functions import deprecated
 from luckydonaldUtils.exceptions import assert_type_or_raise
 
 from ..exceptions import TgApiServerException, TgApiParseException
@@ -142,10 +143,25 @@ class SyncBot(BotBase):
         return self.do(command, **kwargs)
     # end def _do_fileupload
 
+    # compatibility
+
+    @deprecated("The function `bot.send_msg(…) is now named `bot.send_message(…)`.")
     def send_msg(self, *args, **kwargs):
-        """ alias to :func:`send_message` """
+        """ alias to the newer :func:`send_message` """
         return self.send_message(*args, **kwargs)
     # end def send_msg
+
+    @deprecated("The function `bot.kick_chat_member(…) is now named `bot.ban_chat_member(…)`.")
+    def kick_chat_member(self, *args, **kwargs):
+        """ alias to the newer :func:`ban_chat_member` """
+        return self.ban_chat_member(*args, **kwargs)
+    # end def kick_chat_member
+
+    @deprecated("The function `bot.get_chat_members_count(…) is now named `bot.get_chat_member_count(…)`.")
+    def get_chat_members_count(self, *args, **kwargs):
+        """ alias to the newer :func:`get_chat_member_count` """
+        return self.get_chat_member_count(*args, **kwargs)
+    # end def get_chat_members_count
 
     # start of generated functions